Created a camera shaker script
Also corrected a minor spelling mistake in Wall Collision Handler.
The camera shaker has very similar methods to the object shaker, but it's a monobehaviour. Maybe the correct way to do this would be to make the object shaker an interface?
The camera shaker resets the shakes if triggered several times, instead of just ignoring the new shake input.
Problem is that it currently needs to be injected into the wallcollisionhandler script.