Wiggler interpolates based on intensity
Modified the wiggler class so it interpolates between two values for magnitude, total duration, and interval between wiggles, based on a parameter that's passed to the wiggle launcher.
Updated the camera shaker so it takes that param in.
Updated the wall collision handler so it passes that param to the camera shaker based on the intensity of the collision.
Also fixed it because it was max instead of min and that didn't fucking work........ ... . . . .